Add Restic to Backup Software
Restic is an amazing OSS tool that is easy to use if you are already
familiar with Git and can connect to a lot of backup targets.

Local, (S)FTP, HTTP REST, AWS S3, Azure and more are already included in
Restic. A bunch of other targets can be connected by setting up RClone
and using RClone as a backup target.